Final version of py script for simplified state proportions
